requirements.txt
cx_oracle
django-admin-interface
django-crispy-forms